+++ src/interpreter/functions/mod.rs
@@ -1,9 +1,11 @@
+mod anonymous;
 mod contract;
 mod definition;
 mod function;
 mod param;
 mod user_defined;
 
+pub use anonymous::AnonymousFunction;
 pub use contract::ContractFunction;
 pub use definition::{
     AsyncMethod, AsyncProperty, FunctionDef, SyncFunction, SyncMethod, SyncProperty,